@@ -8,12 +8,20 @@ module.exports = {
             {
                 test: /\.js$/,
                 use: [
-                    'babel-loader',
+                    {
+                        loader: 'babel-loader',
+                        options: {
+                            presets: [
+                                "@babel/preset-env",
+                                "@babel/preset-react"
+                            ]
+                        }
+                    }
                 ]
             },
             {
                 test: /\.css$/,
-                loaders: [
+                use: [
                     'style-loader',
                     'css-loader'
                 ],
