@@ -1,9 +1,11 @@
 import * as React from "react";
+import { IJodit } from "jodit";
 
 export interface JoditProps {
     value: string;
-    onChange(value: string): void;
-    config: any;
+    onChange?: (value: string) => void;
+    onBlur?: (value: string) => void;
+    config?: IJodit['options'];
 }
 
 export default class JoditEditor extends React.Component<JoditProps, any> {
