@@ -1,10 +1,13 @@
-var webpack = require("webpack");
-var path = require('path');
+const webpack = require("webpack");
+const path = require('path');
 
-module.exports = {
-    entry: './src/index.js',
-    devtool: "inline-source-map",
-    module: {
+module.exports = (env, argv) => {
+    const debug = !argv || !argv.mode || !argv.mode.match(/production/);
+
+    return {
+        entry: './src/index.js',
+        devtool: debug ? 'inline-source-map' : false,
+        module: {
         rules: [
             {
                 test: /\.js$/,
@@ -14,7 +17,7 @@ module.exports = {
             },
             {
                 test: /\.css$/,
-                loaders: [
+                use: [
                     'style-loader',
                     'css-loader'
                 ],
@@ -22,39 +25,41 @@ module.exports = {
         ]
     },
 
-    resolve: {
-        alias: {
-            "jodit-react": './src'
-        },
-    },
-
-    output: {
-        path: path.join(__dirname, '/build/'),
-        filename: 'jodit-react.js',
-        library: 'JoditEditor',
-        libraryTarget: 'umd'
-    },
-    plugins: [
-        new webpack.DefinePlugin({
-            'process.env': {
-                NODE_ENV: JSON.stringify('production'),
+        resolve: {
+            alias: {
+                "jodit-react": './src'
             },
-        }),
-        new webpack.optimize.ModuleConcatenationPlugin(),
-    ],
-    externals: {
-        jodit: 'jodit',
-        react: {
-            root: 'React',
-            commonjs2: 'react',
-            commonjs: 'react',
-            amd: 'react'
         },
-        'react-dom': {
-            root: 'ReactDOM',
-            commonjs2: 'react-dom',
-            commonjs: 'react-dom',
-            amd: 'react-dom'
+
+        output: {
+            path: path.join(__dirname, '/build/'),
+                filename: 'jodit-react.js',
+                library: 'JoditEditor',
+                libraryTarget: 'umd'
+        },
+        plugins: [
+            new webpack.DefinePlugin({
+                'process.env': {
+                    NODE_ENV: JSON.stringify(debug ? 'development' : 'production'),
+                },
+            }),
+            new webpack.optimize.ModuleConcatenationPlugin(),
+        ],
+        externals: {
+            jodit: 'jodit',
+            Jodit: 'Jodit',
+            react: {
+                root: 'React',
+                    commonjs2: 'react',
+                    commonjs: 'react',
+                    amd: 'react'
+            },
+            'react-dom': {
+                root: 'ReactDOM',
+                    commonjs2: 'react-dom',
+                    commonjs: 'react-dom',
+                    amd: 'react-dom'
+            }
         }
     }
 };
